+++ utils/time.go
@@ -74,3 +74,8 @@ func ParseTime(s string) (time.Time, error) {
 	}
 	return time.Time{}, fmt.Errorf("failed to parse date time string: %s", s)
 }
+
+// GetTime returns the current time.
+func GetTime() time.Time {
+	return time.Now().UTC()
+}