+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## UNRELEASED
+
+- Add Model Inference History to runtime [#186](https://github.com/gohypermode/runtime/pull/186)
+- Pass auth headers correctly when invoking a GraphQL API [#196](https://github.com/gohypermode/runtime/pull/196)
+- Use shared manifest module to read `hypermode.json` [#199](https://github.com/gohypermode/runtime/pull/199)
+
 ## 2024-05-13 - Version 0.7.0
 
 - Sentry is no longer used when `HYPERMODE_DEBUG` is enabled [#187](https://github.com/gohypermode/runtime/pull/187)